Wipe out required use of reverse_bits in main
reverse_bits() was needed since forcecrc32 is operating in the space of
math theory, where the most significant bits are more significant.  CRC32
was created in the space of hardware, where often the first bit shifted
off for checksum was the least significant.  Since there is now the
option to build using libz's crc32() function, reverse_bits() is now
optional in the main module.